Separate out the Model for general 3d model handling. Still some bugs, work in progress
Separate out the Model for general 3d model handling. Still some bugs,
work in progress

+++ trunk/android/StarLib/src/edu/illinois/mitra/starl/motion/MotionAutomaton.java
@@ -136,6 +136,7 @@ public class MotionAutomaton extends RobotMotion {
 		Scanner in = new Scanner(gvh.gps.getMyPosition().name).useDelimiter("[^0-9]+");
 		int index = in.nextInt();
 		ObstacleList obsList = gvh.gps.getViews().elementAt(index);
+		// work in progress here
 		goTo(dest, obsList);
 	}
 
@@ -584,7 +585,7 @@ public class MotionAutomaton extends RobotMotion {
 		int y0 = mypos.y;
 		int x1 = destination.x;
 		int y1 = destination.y;
-		int theta = mypos.angle;
+		int theta = (int)mypos.angle;
 		double alpha = -180 + Math.toDegrees(Math.atan2((y1 - y0), (x1 - x0)));
 		double rad = -(Math.sqrt(Math.pow(x1 - x0, 2) + Math.pow(y1 - y0, 2)) / (2 * Math.sin(Math.toRadians(alpha - theta))));
 		return (int) rad;
